Server modifications to make possible conversation with the frontend

+++ src/main/java/ember/sample/repository/CustomerRepository.java
@@ -1,9 +1,11 @@
 package ember.sample.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import ember.sample.model.Customer;
 
-public interface CustomerRepository extends MongoRepository<Customer, Long> {
+public interface CustomerRepository extends MongoRepository<Customer, BigInteger> {
 
 }
